Possibly Implemented GetPieceOnTile(Tile t)
This is an interesting predicament. I need to see if there is a way to return either a DeckMaster, Card, NPC, or something that implements the IBattlePiece interface, and I want to try that with a method that covers that all, I can account for general methods and type checks elsewhere, in method implementations, if that makes sense. It probably won't when I'm coherent sometime in the future lol :b: I CAN USE EMOJIS WHAT