UI Experiments, Board with Tile Prefabs, more fun stuff!
Working on other things to save my brain from combusting. Getting the UI started was fun, and it helps me actually determine where to call OnMove() from. As well, having the board all on screen looks super nice, as well as having fun with Lerping the positions of the Tiles. I left some runtime tests in :)!