UI DAY :) :) :)
MENU ADVANCEMENT ACHIEVED with the help of refactored functions, a MenuText enum, and switch cases. :)
Today, I've decided that it's UI Day, so I would celebrate by making refactored menu advancement code.
I still have to do different layouts and more practical events(like Tile Selection) for the game, but that's also on the docket for tonight.
It's good to be back!