Modular UI Interface Implented!
I have sown the seeds of a modular menu system that is drawn somewhat efficiently from a specified/altered panel. Everything draws itself from that panel so that wherever it appears, it is drawn correctly relative to its panel. I might need to derive a formula for the scale of these, but who knows. Here's a basic modular menu system so far!

:bee: :bee: :bee: :hamburger: BEESECHURGER?