Just feeling out Events
I know how I'll have to make the event system. From the UI, I'll record the player's choices in their turn, and give an order of evaluation to these. A struct, perhaps? Each of these will have a type and an overloaded method, perform() or something, that calls an event pointing to the User and Target to do things like that. All in EventManager, totally away from BattlePiece and things like that. :)