In-Class Ideas: BattleDetails
I had all of these ideas during a Philosophy class. It's all about whether I needed a static TurnInvoker, and then about implementing a system that saves all of this.
There'll be an abstract BattleDetails class that says what the Board will be like. That, whether the player won, and other info will be stored in a BattleRecord.