More Planning of Animations!
The end is the most useful part. I'm thinking of having a shared BattlePieceAnimations enumeration or something similar to specify generic animations for each BattlePiece that requires animation. I do have to think about some sort of animation pipeline to consistently run the animations that are to be run.

That's all for now, time to be happy and run around like a freak.

:dromedary_camel: :)